style: calendarInput style 수정 및 리팩토링
