+++ src/components/common/Input/CalendarInput/styled/styled.js
@@ -1,12 +1,16 @@
 import styled from 'styled-components';
 
 export const InputContainer = styled.div`
-    width: 633px;
+    width: 490px;
     height: fit-content;
     display: flex;
     flex-direction: column;
     align-items: flex-start;
     gap: 5px;
+
+    @media (max-width: 1024px) {
+        width: 633px;
+    }
     @media (max-width: 431px) {
         width: fit-content;
         height: fit-content;
@@ -24,7 +28,7 @@ export const Label = styled.label`
 `;
 
 export const StyledInputWrapper = styled.div`
-    width: 633px;
+    width: inherit;
     height: 60px;
     display: flex;
     flex-direction: row;
@@ -42,75 +46,5 @@ export const StyledInputWrapper = styled.div`
         width: 340px;
         padding: 10px 15px;
         height: 50px;
-    }
-`;
-
-export const DateInput = styled.div.withConfig({
-    shouldForwardProp: (prop) => prop !== 'isInline',
-})`
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    @media (max-width: 431px) {
-        .react-datepicker {
-            font-size: 12px;
-            width: 180px;
-        }
-
-        .react-datepicker__day {
-            width: 1.2rem;
-            height: 1.2rem;
-            font-size: 12px;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-        }
-        .react-datepicker__header {
-            width: 180px;
-            font-size: 11px;
-        }
-        .react-datepicker__day-names {
-            display: flex;
-            justify-content: space-between;
-        }
-        .react-datepicker__week {
-            display: flex;
-            justify-content: center;
-            align-items: center;
-        }
-    }
-    input {
-        font-size: 16px;
-        width: ${(props) => (props.isInline ? 'auto' : '200px')};
-        border: none;
-        background: ${(props) => (props.isInline ? 'none' : '#ffffff')};
-        padding: ${(props) => (props.isInline ? '0' : '4px 8px')};
-        outline: none;
-        @media (max-width: 431px) {
-            font-size: 10px;
-            width: 100px;
-            padding: 0px 0px 5px 0px;
-        }
-    }
-`;
-
-export const Icon = styled.img`
-    width: ${({ $size }) => $size || '32px'};
-    height: ${({ $size }) => $size || '32px'};
-    cursor: pointer;
-
-    @media (max-width: 431px) {
-        width: 24px;
-        height: 24px;
-    }
-`;
-
-export const DateDivider = styled.span`
-    font-size: 16px;
-    font-weight: 500;
-    color: #d9d9d9;
-    padding-right: 20px;
-    @media (max-width: 431px) {
-        padding-right: 0px;
     }
 `;
